Removing Temp entities and its associated migrations, models, controllers, tests, and dependencies
