+++ app/User.php
@@ -38,10 +38,6 @@ class User extends Authenticatable
         return $this->hasMany(Record::class);
     }
 
-    public function temp_records() {
-        return $this->hasMany(TempRecord::class);
-    }
-
     /**
      * CRUD on other models
      */
@@ -52,9 +48,5 @@ class User extends Authenticatable
 
     public function create_record(Record $record) {
         return $this->records()->save($record);
-    }
-
-    public function create_temp_record(TempRecord $record) {
-        return $this->temp_records()->save($record);
     }
 }
