+++ routes/web.php
@@ -39,7 +39,7 @@ Route::group(['prefix'=>'dashboard', 'middleware' => 'auth'], function () {
     Route::get('/record_issuers/{record_issuer}', 'RecordIssuerController@show')->name('show_record_issuer');
     Route::post('/record_issuers', 'RecordIssuerController@store');
 //    Route::post('/record_issuers/{record_issuer}/records', 'RecordIssuerController@store_record')->name('records');
-    Route::post('/record_issuers/{record_issuer}/temp_records', 'RecordIssuerController@store_temp_record')->name('store_temp_record');
+    Route::post('/record_issuers/{record_issuer}/upload_record', 'RecordIssuerController@upload_record_file')->name('upload_record_file');
     Route::delete('/record_issuers/{record_issuer}', 'RecordIssuerController@destroy');
 
     Route::get('/records/{record}', 'RecordController@show')->name('show_record_file');
@@ -48,11 +48,11 @@ Route::group(['prefix'=>'dashboard', 'middleware' => 'auth'], function () {
     Route::get('/records/{record}/edit', 'RecordController@edit')->name('edit_record'); // show form for edit record
     Route::put('/records/{record}', 'RecordController@update')->name('update_record'); // update record in database
 
-    Route::get('/temp_records/{temp_record}/coords', 'TempRecordController@show_extract_coords_page')->name('show_extract_coords_page');
-    Route::post('/temp_records/{temp_record}/coords', 'TempRecordController@extract_coords')->name('extract_coords');
-    Route::post('/temp_records/{temp_record}/values', 'TempRecordController@confirm_values')->name('confirm_values');
+    Route::get('/records/{record}/coords', 'RecordController@show_extract_coords_page')->name('show_extract_coords_page');
+    Route::post('/records/{record}/coords', 'RecordController@extract_coords')->name('extract_coords');
+    Route::post('/records/{record}/values', 'RecordController@confirm_values')->name('confirm_values');
 
-    Route::get('/temp_record_pages/{temp_record_page}', 'TempRecordPageController@show')->name('show_temp_record_page');
+    Route::get('/record_pages/{record_page}', 'RecordPageController@show')->name('show_record_page');
 });
 
 // stats routes
